+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## NEXT RELEASE
 
+- Catch http error for better message on cli "get" method
 - Add a filtering of closed stations in search results.
 
 ## [0.1.0] - 2026-02-12